requirements.txt
# requirements.txt

requests
gTTS
playsound==1.2.2
streamlit
pymysql
pandas
python-dotenv
pyttsx3
supabase